Shorten displayed buffer names
Using a variation of code written by Matt Tolton for lusty-explorer.vim

Shorten each buffer name by removing all path elements which are not
needed to differentiate a given name from other names.  This usually
results in only the basename shown, but if several buffers of the same
basename are opened, there will be more.