Fix some quoting issues when invoking :LustyFilesystemExplorerStart
As an example, when editing a file inside a directory containing
a single quote, you call :LustyFilesystemExplorerFromHere, the resulting
directory navigated to is stripped of its single quote. This is
generally not where you want to go. A straightforward fix is to avoid
the mess that is vim quoting escapes by pulling the function parameter
straigt into ruby.