Refactors LustyJuggler class for subclassing.
The LustyJuggler class in juggler.rb required refactoring to make it
possible to subclass the class for the purpose of allowing alternative
layouts. A central challenge was the reliance on a class variable
("@@KEYS"). Class variables can not be overridden for only subclasses;
setting "@@KEYS" to a different value in any subclass caused the mapping
to change also for the LustyJuggler parent class. See the following
webpages for further explanation on the complications of inheritance
with class variables

* http://rubysource.com/class-variables-a-ruby-gotcha/
* http://sporkmonger.com/2007/2/19/instance-variables-class-variables-and-inheritance-in-ruby
* http://ghouston.blogspot.com/2006/06/ruby-class-variables-attributes-and.html

This necessitated moving the variable to an instance variable, where it
is still private, but can be overridden easily by subclasses.

Further refactoring involved breaking apart the "@@KEYS" variable into
alphabetic and numeric components, and storing the cancels keys in an
instance variable, too, then looping through these data structures to
map and unmap keys, rather than have a dozen hard-coded mappings and
unmappings.

Overall the result is a LustyJuggler class that can be subclassed
somewhat easily to change the mappings.