Adds support for the Dvorak layout to LustyJuggler.
Remaps the alphabetic keys for LustyJuggler navigation to match the home
row keys for the Dvorak layout. This takes place by subclassing the
"LustyJuggler" class in juggler.rb, given as LustyJugglerDvorak.

The user can designate the desire to use the Dvorak layout for
LustyJuggler by placing the following line in his/her .vimrc:

    let g:LustyJugglerKeyboardLayout = "dvorak"

LustyJuggler will read this configuration variable and set the mapping
to Dvorak if the variable exists and is set to "dvorak".