Change titles of explorer windows to fix odd behavior in Vim 7.3
From e.g. "[LustyExplorer-Files]" to "LustyExplorer--Files".  This is
because internally we call something like `:split [LustyExplorer-Files]`
in order to create the explorer window.  The argument to :split is the
name of the file to edit (or, if it doesn't exist, a new buffer with
that name).  However, :split does some wildcard expansion on its
argument, and can interpret [LustyExplorer-Files] as a regex-ish range.

In most versions of Vim, this :split call may emit a warning, but Vim
will go ahead and create the new buffer named [LustyExplorer-Files] in a
new window.  In some versions of Vim 7.3, however, it will emit the
warning (stifled by `!silent` in lusty-explorer.vim) and not create the
new buffer or the new window.

I could probably work around this by escaping the square bracket
characters, but I'm not sure how that will affect Windows, in which the
escape character '\' is also a directory separator.  So, I'll do the
easier thing and use more innocuous names for the explorer windows.