lustyjuggler: preserve the "<script>" tag when restoring a mapping
This is a true fix for #20.  The `maparg()` function still does not
report back enough information to know whether or not the "<script>" tag
was specified for the mapping.  So, let's take another approach and
parse the output of the map command.  The presence of an "&" appears to
be a good indicator that you need the "<script>" tag.  Let's key off
that and use that information when generating the command to restore the
mapping.