[Fixes #103]: LustyJuggler: don't splash mapping output to the editor on launch
LustyJuggler sometimes uses Vim's :redir on launch as part of figuring
out how to restore mappings that it temporarily changes.  It turns out
that :redir tees its output to the message area in addition to the place
it's redirecting to, and for some folks' configurations, that causes a
mess, up to and including the bogus "Press ENTER or type command to
continue".

We can avoid the tee by wrapping the :redir call in a Vim function
called with :silent.  Really ugly, but it seems to work.